docs(progress): update learnings for path hijacking fix
Documented technical insights and anti-patterns for the path
hijacking security fix in progress/LEARNINGS.md.
