+++ .gitignore
@@ -92,4 +92,3 @@ csm_test*
 .claude/settings.local.json
 .claude/skills
 *.tgz
-memory_retention_curve.csv
