+++ .ignore
@@ -2,4 +2,3 @@
 !.tmp/
 !*.tmp
 !*.tgz
-!memory_retention_curve.csv
