+++ CHANGELOG.md
@@ -5,26 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.4] - 2026-04-24
-
-### Fixed
-
-- **NEON SIMD Build**: Fixed macOS arm64 build by:
-  - Using correct `vld1q_u64`/`vst1q_u64` pointer types (`*const u64` not `*const uint64x2_t`)
-  - Adding explicit unsafe block required by `#[target_feature(enable = "neon")]`
-- **Release Workflow**: Added `make_latest: true` to ensure new releases are marked as latest
-
-### Changed
-
-- **Benchmark Verification**: Updated metrics showing significant improvements:
-  - `reservoir_step_50k`: 76.6µs → 55.7µs (44% faster)
-  - `batch_similarity_1000`: 470µs → 195µs (58% faster)
-- **GOAP State**: Wave 21 resolved, npm OIDC publishing verified working
-
-## [0.3.3] - 2026-04-24 [YANKED]
-
-This version was tagged but contained a build failure on macOS arm64. Use v0.3.4 instead.
-
 ## [0.3.2] - 2026-04-09
 
 ### Added
@@ -115,20 +95,6 @@ This version was tagged but contained a build failure on macOS arm64. Use v0.3.4
 - **WASM size gate**: Fixed script to check library WASM instead of CLI binary.
 
 ## [Unreleased]
-
-### Added
-
-- **PR #102**: Optimized HVec10240::permute by eliminating modulo and branching (~15% faster)
-- **PR #100**: Phase 60 test suites and benchmarks for comprehensive coverage
-- **ADR-0064**: Inertial Reservoir Dynamics (InertialESN) - second-order momentum for improved
-  temporal memory retention. Beta parameter [0.0, 0.5] for momentum control.
-- **ADR-0065**: Selectivity-Aware Filtered Retrieval - strategy routing based on filter
-  selectivity ratio (PreFilter, BucketPostFilter, ScanPostFilter).
-
-### Changed
-
-- **GOAP Audit**: Resolved 3 stale queued actions (test_inertial_reservoir, test_selectivity_aware_retrieval,
-  benchmark_inertial_reservoir) - work already complete.
 
 ## [0.2.9] - 2026-04-06
