+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "chaotic_semantic_memory"
-version = "0.3.4"
+version = "0.3.2"
 edition = "2024"
 rust-version = "1.85"
 description = "AI memory systems with hyperdimensional vectors and chaotic reservoirs"
