+++ wasm/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@d-o-hub/chaotic_semantic_memory",
-  "version": "0.3.4",
+  "version": "0.3.2",
   "description": "AI memory systems with hyperdimensional vectors and chaotic reservoirs - WASM bindings",
   "main": "chaotic_semantic_memory.js",
   "types": "chaotic_semantic_memory.d.ts",
